+++ examples/hex_maze_image.rs
@@ -1,3 +1,5 @@
+#![allow(clippy::unwrap_used)]
+#![allow(clippy::expect_used)]
 use std::fs;
 
 use bevy_knossos::maze::{Cell, HexMazeBuilder, RecursiveBacktracking};