+++ examples/mazes.rs
@@ -1,3 +1,5 @@
+#![allow(clippy::unwrap_used)]
+#![allow(clippy::expect_used)]
 /// Extensive set of examples on how to use the lib to generate mazes and save them to files
 use bevy_knossos::maze::{
     AldousBroder, AsciiBroad, Bias, BinaryTree, Eller, GameMap, GrowingTree, HuntAndKill, Image,