+++ src/maze/grid/cell.rs
@@ -121,6 +121,7 @@ impl CellStatus {
     }
 }
 #[cfg(test)]
+#[allow(clippy::unwrap_used)]
 mod tests {
     use super::Cell;