+++ src/pathfind.rs
@@ -111,6 +111,7 @@ pub(crate) fn find_path(
 }
 
 #[cfg(test)]
+#[allow(clippy::unwrap_used)]
 mod tests {
     use super::*;