+++ src/utils/arena.rs
@@ -50,6 +50,7 @@ impl ArenaTree {
 impl Node {}
 
 #[cfg(test)]
+#[allow(clippy::unwrap_used)]
 mod tests {
     use super::*;